requirements.txt
crewai>=0.11.0
python-dotenv>=1.0.0
Pillow>=10.0.0
PyPDF2>=3.0.1
PyMuPDF>=1.23.0